Fix error and bugs

+++ services/commentService.js
@@ -5,7 +5,9 @@ const {
   STATUSCODE,
   RESOURCE
 } = require("../constants/index");
-const { cloudinary } = require("../utils/cloudinary");
+const {
+  cloudinary
+} = require("../utils/cloudinary");
 
 exports.getAllCommentData = async () => {
   const comments = await Comment.find().sort({
@@ -71,8 +73,9 @@ exports.updateCommentData = async (req, res, id) => {
   const existingComment = await Comment.findById(id).lean().exec();
 
   let images = existingComment.image || [];
+
   if (req.files && Array.isArray(req.files) && req.files.length > 0) {
-    images = await Promise.all(
+    const newImages = await Promise.all(
       req.files.map(async (file) => {
         const result = await cloudinary.uploader.upload(file.path, {
           public_id: file.filename,
@@ -85,18 +88,24 @@ exports.updateCommentData = async (req, res, id) => {
       })
     );
 
-    await cloudinary.api.delete_resources(
-      existingComment.image.map((image) => image.public_id)
-    );
+    images = [...images, ...newImages];
+
+    if (existingComment.image && existingComment.image.length > 0) {
+      await cloudinary.api.delete_resources(
+        existingComment.image.map((image) => image.public_id)
+      );
+    }
   }
 
-  const updatedComment = await Comment.findByIdAndUpdate(id, {
-      ...req.body,
-      image: images,
-    }, {
-      new: true,
-      runValidators: true,
-    })
+  const updatedComment = await Comment.findByIdAndUpdate(
+      id, {
+        ...req.body,
+        image: images,
+      }, {
+        new: true,
+        runValidators: true,
+      }
+    )
     .lean()
     .exec();
 
@@ -108,6 +117,7 @@ exports.updateCommentData = async (req, res, id) => {
 
   return updatedComment;
 };
+
 
 exports.deleteCommentData = async (id) => {
   if (!mongoose.Types.ObjectId.isValid(id))
@@ -122,7 +132,9 @@ exports.deleteCommentData = async (id) => {
 
   if (publicIds.length > 0) await cloudinary.api.delete_resources(publicIds);
 
-  await Comment.deleteOne({ _id: id }).lean().exec();
+  await Comment.deleteOne({
+    _id: id
+  }).lean().exec();
 
   return comment;
 };
